frontend dir and OrganizationService deleted, added ActivitiesService, PlansService, RepositoryService, UserService, OrganizationRepository, OrganizationRepositoryDao and OrganizationRepositoryDaoImpl
