+++ App/src/main/java/com/korpodrony/model/ActivitiesType.java
