+++ App/src/main/java/com/korpodrony/model/Activity.java
