+++ App/src/main/java/com/korpodrony/model/Plan.java
