+++ App/src/main/java/com/korpodrony/service/PropertiesService.java
