+++ App/src/main/java/com/korpodrony/utils/IoTools.java
