+++ App/src/main/java/com/korpodrony/utils/JSONReader.java
