--- frontend/index.html
@@ -1,96 +1,0 @@
-<!doctype html>
-<html lang="pl">
-
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <link rel="stylesheet" href="style.css">
-
-    <title>Korpodrony</title>
-</head>
-
-<body>
-    <div id="particles-js">
-        <p class="banner">Korpodrony</p>
-
-        <nav class="navbar navbar-expand-lg navbar-light bg-primary">
-            <a class="navbar-brand col-1"></a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav col-10">
-                    <li class="nav-item active">
-                        <a class="nav-link text-white" href="users.html">Użytkownicy <span class="sr-only">(current)</span></a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="activities.html">Zajęcia</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="plans.html">Plany</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white ml-auto" href="search.html">Szukaj</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-
-        <div class="container pr-5 pl-5 mt-5 mb-5">
-            <div class="row">
-                <div col-6 class="ml-auto m-5">
-                    <div class="card" style="width: 18rem;">
-                        <p class="m-3">PATRYK</p>
-                        <img src="pobrane.jpg" class="card-img-top" alt="...">
-                        <div class="card-body">
-                            <p class="card-text">Szkielet frontendowy aplikacji</p>
-                        </div>
-                    </div>
-                </div>
-                <div col-6 class="mr-auto m-5">
-                    <div class="card" style="width: 18rem;">
-                        <p class="m-3">Konrad</p>
-                        <img src="misterion.png" class="card-img-top" alt="...">
-                        <div class="card-body">
-                            <p class="card-text">...</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div col-6 class="ml-auto m-5">
-                    <div class="card" style="width: 18rem;">
-                        <p class="m-3">Zuza</p>
-                        <img src="misterion.png" class="card-img-top" alt="...">
-                        <div class="card-body">
-                            <p class="card-text">...</p>
-                        </div>
-                    </div>
-                </div>
-                <div col-6 class="mr-auto m-5">
-                    <div class="card" style="width: 18rem;">
-                        <p class="m-3">Błażej</p>
-                        <img src="misterion.png" class="card-img-top" alt="...">
-                        <div class="card-body">
-                            <p class="card-text">...</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <p class="m-3"> Nasz projekt to aplikacja, napiszcie tutaj co to ma robić... :)</p>
-            </div>
-        </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-    <script src="particles.min.js"></script>
-    <script src="app.js"></script>
-</body></html>
