--- frontend/style.css
@@ -1,31 +1,0 @@
-canvas{
-    width: 500px !imporant;
-    position: fixed;
-    top: 0;
-    background-color: #3c3c3c;
-    z-index: -500 !important;
-}
-
-.container{
-    z-index: 100;
-    position: relative;
-    background-color: #fff;
-    box-sizing: border-box;
-    border-radius: 1rem;
-}
-
-.banner{
-    display: block;
-    height: 7rem;
-    color: #FFf;
-    font-size: 4rem;
-    text-align: center;
-    background-color: rgb(156, 159, 255, 0.3);
-    margin: 0;
-    box-sizing: border-box;
-    font-style: italic;
-}
-
-.btn-group{
-    display: flex;
-}
