--- frontend/user.html
@@ -1,102 +1,0 @@
-<!doctype html>
-<html lang="pl">
-
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <link rel="stylesheet" href="style.css">
-
-    <title>Korpodrony</title>
-</head>
-
-<body>
-    <div id="particles-js">
-        <p class="banner">Korpodrony</p>
-
-        <nav class="navbar navbar-expand-lg navbar-light bg-primary">
-            <a class="navbar-brand col-1"></a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav col-10">
-                    <li class="nav-item active">
-                        <a class="nav-link text-white" href="activities.html">Użytkownicy</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="activities.html">Zajęcia</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="plans.html">Plany</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white ml-auto" href="search.html">Szukaj</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-
-        <div class="container pr-5 pl-5 mt-5 mb-5">
-
-            <div class="row">
-                <div class="col-10 ml-auto mr-auto mt-5">
-                    <h2>Użytkownik</h2>
-                    <form>
-                        <div class="form-group row">
-                            <label for="staticEmail" class="col-sm-2 col-form-label">ID:</label>
-                            <div class="col-sm-10">
-                                <input type="text" readonly class="form-control-plaintext" id="staticEmail" value="1">
-                            </div>
-                        </div>
-                        <div class="form-group row">
-                            <label for="inputPassword" class="col-sm-2 col-form-label">Imię</label>
-                            <div class="col-sm-10">
-                                <input type="password" class="form-control" id="inputName" placeholder="Imię">
-                            </div>
-                        </div>
-                        <div class="form-group row">
-                            <label for="inputPassword" class="col-sm-2 col-form-label">Nazwisko</label>
-                            <div class="col-sm-10">
-                                <input type="password" class="form-control" id="inputSurname" placeholder="Nazwisko">
-                            </div>
-                        </div>
-                    </form>
-                </div>
-            </div>
-
-            <div class="row mt-3">
-                <div class="col-3 mb-3 ml-auto">
-                    <a class="btn btn-success text-white w-100">Zmień</a>
-                </div>
-                <div class="col-3 mb-3 mr-auto">
-                    <a class="btn btn-danger text-white w-100">usuń</a>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-10 ml-auto mr-auto mt-5">
-                    <h2>Zajęcia, do których przypisany jest użytkownik</h2>
-                    <ul class="list-group w-100 mb-3">
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="activity.html" class="stretched-link">Go dadadsa</a></li>
-                    </ul>
-                </div>
-            </div>
-        </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-    <script src="particles.min.js"></script>
-    <script src="app.js"></script>
-</body></html>
