--- frontend/users.html
@@ -1,77 +1,0 @@
-<!doctype html>
-<html lang="pl">
-
-<head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
-    <link rel="stylesheet" href="style.css">
-
-    <title>Korpodrony</title>
-</head>
-
-<body>
-    <div id="particles-js">
-        <p class="banner">Korpodrony</p>
-
-        <nav class="navbar navbar-expand-lg navbar-light bg-primary">
-            <a class="navbar-brand col-1"></a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav col-10">
-                    <li class="nav-item active">
-                        <a class="nav-link">Użytkownicy</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="activities.html">Zajęcia</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white" href="plans.html">Plany</a>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link text-white ml-auto" href="search.html">Szukaj</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-
-        <div class="container pr-5 pl-5 mt-5 mb-5">
-
-            <div class="row">
-                <div class="col-10 ml-auto mr-auto mt-5">
-                    <h2>Użytkowicy</h2>
-                    <ul class="list-group w-100 ">
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                        <li class="list-group-item"><a href="user.html" class="stretched-link">Go dadadsa</a></li>
-                    </ul>
-                </div>
-            </div>
-            <div class="row mt-3">
-                <div class="col-3 ml-auto mr-auto mb-3">
-                    <a class="btn btn-success text-white w-100">Dodaj</a>
-                </div>
-            </div>
-
-        </div>
-    </div>
-
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-    <script src="particles.min.js"></script>
-    <script src="app.js"></script>
-</body>
-
-</html>
